url-service/requirements.txt
fastapi
uvicorn
torch
torch-geometric
pydantic
python-dotenv
scikit-learn
tldextract
